Move usersAnalysis to pipeline dir
